git-svn-id: http://svn.code.sf.net/p/rts-2/code/branches/rts-2/REL_0_8_0@5559 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/grb/rts2conngrb.h
@@ -38,8 +38,8 @@ class Rts2ConnGrb:public Rts2ConnNoSend
 
 		int gcnReceivedBytes;	 // number of bytes received
 
-		long lbuf[SIZ_PKT];		 // local buffer - swaped for Linux
-		long nbuf[SIZ_PKT];		 // network buffer
+		int32_t lbuf[SIZ_PKT];		 // local buffer - swaped for Linux
+		int32_t nbuf[SIZ_PKT];		 // network buffer
 		struct timeval last_packet;
 		double here_sod;		 // machine SOD (seconds after 0 GMT)
 		double last_imalive_sod; // SOD of the previous imalive packet
